+++ chocolatey/winbox/3.37/tools/chocolateyInstall.ps1
@@ -1,0 +1,45 @@
+﻿$packageName = 'winbox'
+$appName = 'WinBox'
+$url32 = 'https://mt.lv/winbox'
+$url64 = 'https://mt.lv/winbox64'
+$dir = $(Split-Path -parent $MyInvocation.MyCommand.Definition)
+$is64 = ((Get-ProcessorBits 64) -and !$env:chocolateyForceX86)
+$processName32 = 'winbox'
+$processName64 = 'winbox64'
+$processName = If ($is64) { $processName64 } Else { $processName32 }
+$exe = $processName + '.exe'
+$fullPath = Join-Path -Path $dir -ChildPath $exe
+$checksum32 = 'b9e0f5a8c21ee6b041e461dae27e92ddeea34a71a84c855aac6ae435e8ad4aa0dfcdae8c211637e8bdf64c752d266d935003e9c78aa3efa55a861e0bd8676357'
+$checksum64 = '01381371f802181e126534510c05ca25e393df7151a4ad3760567c651ac44ec34748c00950f6ec6ee9865ce97195e4cc7f1a55e85a36b54f611dd8c9aae63519'
+$checksumType = 'sha512'
+
+$startFolder = $appName
+
+Function CreateStartMenuShortcut {
+    $startMenuFolderPath = Join-Path -Path ([Environment]::GetFolderPath('Programs')) `
+        -ChildPath $startFolder
+    New-Item $startMenuFolderPath -Type Directory -ErrorAction SilentlyContinue
+    If (Test-Path -Path $startMenuFolderPath -PathType Container) {
+            Install-ChocolateyShortcut -shortcutFilePath `
+                (Join-Path -Path $startMenuFolderPath -ChildPath ($appName + '.lnk')) `
+                -targetPath $fullPath -workingDirectory $dir
+    } Else {
+        Write-Warning "Start menu folder couldn't be created."
+    }
+}
+
+# SCRIPT STARTS HERE #
+
+Stop-Process -Name $processName32 -ErrorAction SilentlyContinue
+Stop-Process -Name $processName64 -ErrorAction SilentlyContinue
+
+# Workaround for https://github.com/chocolatey/choco/issues/952
+Remove-Item -Path $fullPath -Force -Confirm:$FALSE -ErrorAction SilentlyContinue
+
+Get-ChocolateyWebFile -PackageName $packageName -FileFullPath $fullPath -Url $url32 -Url64bit $url64 -Checksum $checksum32 -ChecksumType $checksumType -Checksum64 $checksum64 -ChecksumType64 $ChecksumType
+
+# create .gui file
+New-Item -Path $dir -Name ($exe + '.gui') -ItemType file
+
+CreateStartMenuShortcut
+