+++ chocolatey/winbox/3.37/tools/chocolateyInstall.ps1.php
@@ -1,0 +1,46 @@
+﻿<?php require '../vars.inc'; ?>
+$packageName = 'winbox'
+$appName = 'WinBox'
+$url32 = '<?php echo URL32; ?>'
+$url64 = '<?php echo URL64; ?>'
+$dir = $(Split-Path -parent $MyInvocation.MyCommand.Definition)
+$is64 = ((Get-ProcessorBits 64) -and !$env:chocolateyForceX86)
+$processName32 = '<?php echo PROCESS_NAME32; ?>'
+$processName64 = '<?php echo PROCESS_NAME64; ?>'
+$processName = If ($is64) { $processName64 } Else { $processName32 }
+$exe = $processName + '.exe'
+$fullPath = Join-Path -Path $dir -ChildPath $exe
+$checksum32 = '<?php echo CHECKSUM32; ?>'
+$checksum64 = '<?php echo CHECKSUM64; ?>'
+$checksumType = 'sha512'
+
+$startFolder = $appName
+
+Function CreateStartMenuShortcut {
+    $startMenuFolderPath = Join-Path -Path ([Environment]::GetFolderPath('Programs')) `
+        -ChildPath $startFolder
+    New-Item $startMenuFolderPath -Type Directory -ErrorAction SilentlyContinue
+    If (Test-Path -Path $startMenuFolderPath -PathType Container) {
+            Install-ChocolateyShortcut -shortcutFilePath `
+                (Join-Path -Path $startMenuFolderPath -ChildPath ($appName + '.lnk')) `
+                -targetPath $fullPath -workingDirectory $dir
+    } Else {
+        Write-Warning "Start menu folder couldn't be created."
+    }
+}
+
+# SCRIPT STARTS HERE #
+
+Stop-Process -Name $processName32 -ErrorAction SilentlyContinue
+Stop-Process -Name $processName64 -ErrorAction SilentlyContinue
+
+# Workaround for https://github.com/chocolatey/choco/issues/952
+Remove-Item -Path $fullPath -Force -Confirm:$FALSE -ErrorAction SilentlyContinue
+
+Get-ChocolateyWebFile -PackageName $packageName -FileFullPath $fullPath -Url $url32 -Url64bit $url64 -Checksum $checksum32 -ChecksumType $checksumType -Checksum64 $checksum64 -ChecksumType64 $ChecksumType
+
+# create .gui file
+New-Item -Path $dir -Name ($exe + '.gui') -ItemType file
+
+CreateStartMenuShortcut
+